LIB: introduce Arch trait
Each architecture has to implement Arch trait.

+++ kernel/src/arch/rv32/config.rs
@@ -1,3 +1,3 @@
-pub const HEAP_START: u64 = 0x8800_0000;
+pub const HEAP_START: usize = 0x8800_0000;
 
 pub const PREEMPT_IRQ: u16 = 0;
